Add toString function to action creators (#98)
This allows people migrating from redux-actions (and using `==` to coerce
those actions to stringst to check against `action.type`) to continue to do
so as part of an incremental transition.

---

Thank you for your contribution! :thumbsup:

Please makes sure that these checkboxes are checked before submitting your PR, thank you!

* [X] Rebase before creating a PR to keep commit history clean
* [X] Clear node_modules and reinstall all the dependencies: `npm run reinstall`
* [X] Run checker npm script: `npm run check`

Extra checklist:
* [X] Always add some description and refer to all the related issues using `#` tag

For new feature:
 * [x] Update API docs
 * [ ] Add examples to demonstrate new feature
 * [ ] Add type unit tests
 * [x] Add runtime unit tests

Left out the above because my intention wasn't necessarily to introduce this as a usable API but
as a js -> ts migration path.